extract MobXHooksObserverComponent
