+++ experiments/mobx_hooks/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:jaspr/jaspr.dart';
 import './app.dart';
-import 'mobx_hooks/mobx_hooks.dart';
+import 'mobx_hooks/jaspr_observer.dart';
 
 void main() {
   runApp(const MobXHooksObserverComponent(child: App()));
