+++ experiments/mobx_hooks/web/main.dart
@@ -1,6 +1,6 @@
 import 'package:jaspr/jaspr.dart';
 import 'package:mobx_hooks_experiment/app.dart';
-import 'package:mobx_hooks_experiment/mobx_hooks/mobx_hooks.dart';
+import 'package:mobx_hooks_experiment/mobx_hooks/jaspr_observer.dart';
 
 void main() {
   runApp(const MobXHooksObserverComponent(child: App()));
